requirements.txt
streamlit
chromadb
google-genai
python-dotenv